docs(dashboard): write down the component and hook conventions
Adds AGENTS.md, and brings the two hooks and the header in line with it: hooks
annotate their return type, and the sign-out button owns its own behaviour
instead of living inside the header.
