+++ apps/dashboard/src/lib/hooks/use-session.ts
@@ -1,5 +1,6 @@
+import type { Session } from '#lib/auth.ts';
 import { Route } from '#routes/__root.tsx';
 
-export function useSession() {
+export function useSession(): Session | null {
   return Route.useRouteContext().session;
 }
